PR: updates
